Backend/SafeLink_Backend/requirements.txt
scapy>=2.5.0
pandas>=2.2.0
numpy>=1.26.0
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
tqdm>=4.66.0
pytest>=8.2.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Threat Intelligence & Network Integration
aiohttp>=3.9.0
pysnmp>=4.4.12

# Task Queue (for automated mitigation)
celery>=5.3.0
redis>=5.0.0

# Additional utilities
websockets>=12.0

# Optional: Advanced ML features
# Uncomment to enable SMOTE balancing and hyperparameter optimization
# imbalanced-learn>=0.12.0
# optuna>=3.5.0
